+++ env/code/main.js
@@ -1,2 +1,1 @@
 addSprite("mark");
-throw new Error("123");