+++ CHANGELOG.md
@@ -11,6 +11,7 @@ k.add(...);
 k.keyPress(...);
 ```
 - audio is now paused when you leave the tab
+- audio is now paused on `debug.paused = true`
 - added localStorage helper `getData(key, default?)` and `setData(key, data)`
 - added `loadShader(id, vert, frag, isUrl)`
 - added `shader()` comp for attaching custom shader to an obj